requirements_old.txt
amqp==5.1.1
apturl==0.5.2
arabic-reshaper==2.1.4
asana==2.0.0
asgiref==3.5.2
asn1crypto==1.5.1
asttokens==2.0.8
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.11.1
billiard==3.6.4.0
blinker==1.4
blis==0.7.8
Brlapi==0.8.3
bs4==0.0.1
cached-property==1.5.2
cachetools==5.2.0
catalogue==2.0.8
celery==5.0.5
certifi==2022.9.24
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.1
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
command-not-found==0.3
commonmark==0.9.1
confection==0.0.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.8
cssselect2==0.7.0
cupshelpers==1.0
cymem==2.0.6
dbus-python==1.2.18
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
Deprecated==1.2.13
deprecation==2.1.0
dictdiffer==0.9.0
discord-webhook==0.17.0
distlib==0.3.6
distro==1.7.0
distro-info===1.1build1
dj-database-url==1.0.0
Django==3.2.12
django-3-jet==1.0.8
django-admin-list-filter-dropdown==1.0.3
django-admin-numeric-filter==0.1.6
django-admin-rangefilter==0.9.0
django-admin-searchable-dropdown==1.2
django-autofixture==0.12.1
django-autoslug==1.9.8
django-braces==1.15.0
django-classy-tags==3.0.1
django-colorfield==0.7.2
django-constance==2.9.1
django-cors-headers==3.13.0
django-countries==7.4.2
django-crispy-forms==1.14.0
django-crontab==0.7.1
django-cuser==2017.3.16
django-environ==0.4.5
django-extensions==3.1.5
django-filter==22.1
django-jsonfield==1.4.1
django-map-widgets==0.4.0
django-oauth-toolkit==2.2.0
django-oauth2-provider==0.2.6.1
django-recaptcha==3.0.0
django-rest-framework-social-oauth2==1.1.0
django-rest-knox==4.2.0
django-rest-passwordreset==1.3.0
django-rest-swagger==2.2.0
django-restframework==0.0.1
django-sekizai==2.0.0
django-six==1.0.4
django-smtp-ssl==1.0
django-social-auth==0.7.1
django-storages==1.13.1
django-tagconstants==0.1
django-templated-mail==1.1.1
django-tinymce==3.5.0
django-utils-six==2.0
djangorestframework==3.14.0
djangorestframework-api-key==2.2.0
djangorestframework-gis==1.0
djangorestframework-simplejwt==4.8.0
djongo==1.3.6
djoser==2.1.0
dnspython==2.2.1
drf-yasg==1.21.4
duplicity==0.8.21
elastic-transport==8.4.0
elasticsearch==8.4.3
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.4.0/en_core_web_sm-3.4.0-py3-none-any.whl
executing==1.1.1
fasteners==0.14.1
filelock==3.8.0
flake8==5.0.4
flake8-string-format==0.3.0
Flask==2.2.2
Flask-SQLAlchemy==3.0.2
fpdf==1.7.2
future==0.18.2
GDAL==3.4.1
geographiclib==2.0
geopy==2.3.0
google-api-core==2.10.2
google-api-python-client==2.65.0
google-auth==2.13.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.7.0
googleapis-common-protos==1.56.4
greenlet==1.1.3.post0
gunicorn==20.0.4
html5lib==1.1
httplib2==0.20.2
huey==2.4.4
idna==2.10
img2pdf==0.4.4
importlib-metadata==4.6.4
inflection==0.5.1
instagram-scraper==1.11.0
ipython==8.5.0
itsdangerous==2.1.2
itypes==1.2.0
jedi==0.18.1
jeepney==0.7.1
Jinja2==3.1.2
joblib==1.2.0
jsonfield==3.1.0
jwcrypto==1.4.2
keyring==23.5.0
kombu==5.2.4
langcodes==3.3.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
lxml==4.9.1
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mccabe==0.7.0
modeltranslation==0.25
monotonic==1.6
more-itertools==8.10.0
murmurhash==1.0.8
netifaces==0.11.0
nltk==3.7
notion==0.0.28
numpy==1.21.4
oauth2==1.9.0.post1
oauth2client==4.1.3
oauthlib==3.2.2
olefile==0.46
openapi-codec==1.3.2
oscrypto==1.3.0
packaging==21.3
paramiko==2.9.3
parso==0.8.3
pathy==0.6.2
pdf2image==1.16.0
pdfrw==0.4
pexpect==4.8.0
pickleshare==0.7.5
pikepdf==6.2.1
Pillow==9.0.1
platformdirs==2.5.2
preshed==3.0.7
prompt-toolkit==3.0.31
protobuf==4.21.9
provider==0.1.0
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.20.1
pycodestyle==2.9.1
pycparser==2.21
pycups==2.0.1
pydantic==1.9.2
pyflakes==2.5.0
Pygments==2.13.0
PyGObject==3.42.1
pyHanko==0.15.0
pyhanko-certvalidator==0.19.5
PyJWT==2.3.0
pymacaroons==0.13.0
pymongo==3.12.3
PyMySQL==0.9.3
PyNaCl==1.5.0
pyotp==2.6.0
pyparsing==3.0.9
PyPDF2==2.11.1
PyPDF3==1.0.6
pyRFC3339==1.1
python-apt==2.3.0+ubuntu2.1
python-bidi==0.4.2
python-dateutil==2.8.1
python-debian===0.1.43ubuntu1
python-dotenv==0.21.0
python-http-client==3.3.7
python-memcached==1.59
python-openid==2.2.5
python-slugify==6.1.2
python3-openid==3.2.0
pytz==2022.4
pytz-deprecation-shim==0.1.0.post0
pyxdg==0.27
PyYAML==5.4.1
qrcode==7.3.1
redis==3.5.3
regex==2022.9.13
reportlab==3.6.8
requests==2.24.0
requests-mock==1.10.0
requests-oauthlib==1.3.1
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
SecretStorage==3.3.1
sendgrid==6.9.7
sentry-sdk==0.16.4
shortuuid==1.0.9
simplejson==3.17.6
six==1.16.0
SkPy==0.10.4
smart-open==5.2.1
social-auth-app-django==4.0.0
social-auth-core==4.3.0
soupsieve==2.3.2.post1
spacy==3.4.1
spacy-legacy==3.0.10
spacy-loggers==1.0.3
SQLAlchemy==1.4.42
sqlparse==0.2.4
srsly==2.4.4
stack-data==0.5.1
starkbank-ecdsa==2.2.0
stripe==5.0.0
svglib==1.4.1
systemd-python==234
text-unidecode==1.3
textblob==0.17.1
thinc==8.1.4
tinycss2==1.2.1
tqdm==4.64.1
traitlets==5.4.0
twilio==7.15.1
typer==0.4.2
typing_extensions==4.4.0
tzdata==2022.5
tzlocal==4.2
ua-parser==0.10.0
ubuntu-advantage-tools==27.12
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==4.1.1
uritools==4.0.0
urllib3==1.25.11
usb-creator==0.3.7
user-agents==2.1
users==1.0.dev0
vat-format-checker==0.0.4
vine==5.0.0
virtualenv==20.16.5
wadllib==1.3.6
wasabi==0.10.1
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.2.2
whitenoise==5.2.0
wrapt==1.14.1
xdg==5
xhtml2pdf==0.2.8
xkit==0.0.0
zipp==1.0.0
